+++ src/config/filehelper.php
@@ -2,4 +2,6 @@
 
 return [
     'max_upload_filesize' => '2000000', // 20,00,000 20 lakh bytes = 2Mb
+    'username' => 'fh_admin',
+    'password' => 'fh_admin@246',
 ];